+++ Source/Plugins/EbsdImport/EbsdImport.h
@@ -75,8 +75,8 @@ class EbsdImport : public AbstractPipeline, public Observer
     virtual ~EbsdImport();
 
     MXA_INSTANCE_STRING_PROPERTY(OutputFile)
-    MXA_INSTANCE_PROPERTY(int, ZStartIndex)
-    MXA_INSTANCE_PROPERTY(int, ZEndIndex)
+    MXA_INSTANCE_PROPERTY(int64_t, ZStartIndex)
+    MXA_INSTANCE_PROPERTY(int64_t, ZEndIndex)
     MXA_INSTANCE_PROPERTY(float, ZResolution)
     MXA_INSTANCE_PROPERTY(std::vector<std::string>, EbsdFileList);
     MXA_INSTANCE_PROPERTY(bool, ReorderArray)
